backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
celery[redis]==5.3.6
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pdfplumber==0.10.4
sentence-transformers==2.3.1
scikit-learn==1.4.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
